backend/requirements.txt
fastapi
uvicorn
diffusers
transformers
torch
pillow
numpy